requirements.txt
face_recognition~=1.3.0
numpy~=1.22.1
pillow~=8.4.0